Match improvements pt 8 (#36)
Update state processor

+++ src/core/src/league/schedule/round.rs
@@ -107,7 +107,7 @@ fn generate_game_pairs(teams: &[u32], tours_count: usize) -> Vec<(u32, u32)> {
     let mut result = Vec::with_capacity(tours_count);
 
     let team_len = teams.len() as u32;
-    let team_len_half = team_len / 2 as u32;
+    let team_len_half = team_len / 2u32;
 
     let mut temp_vec = Vec::with_capacity((team_len_half + 1) as usize);
